chore: factorize common {Triplet | Center}Loss code
